specs: emit better parsing errors for specs.
Parse error information is kept for specs, but it doesn't seem like we propagate it
to the user when we encounter an error.  This fixes that.

e.g., for this error in a package:

```python
    depends_on("python@:3.8", when="0.900:")
```

Before, with no context and no clue that it's even from a particular spec:

```
==> Error: Unexpected token: ':'
```

With this PR:

```
==> Error: Unexpected token: ':'
  Encountered when parsing spec:
    0.900:
         ^
```

I would like to show context like we do with actual build-time package errors,
but that's more work because it seems like the import mechanism strips package
frames off the stack.  We can try to preserve context like that from directives,
but it'll take more implemetnation and I am not sure how much it'll cost in terms
of package import time. This is at least an improvement.